fix code smells

+++ unobstructed_lanechange/include/unobstructed_lanechange.h
@@ -120,8 +120,8 @@ namespace unobstructed_lanechange
             double minimum_lookahead_speed_ = 2.8;
             double maximum_lookahead_speed_ =13.9;
             double lateral_accel_limit_ = 1.5;
-            double speed_moving_average_window_size_ = 5;
-            double curvature_moving_average_window_size_ = 9;
+            int speed_moving_average_window_size_ = 5;
+            int curvature_moving_average_window_size_ = 9;
             double curvature_calc_lookahead_count_ = 1;
             int downsample_ratio_ =8;
             bool enable_object_avoidance_lc_ = false;
